# backend/src/docs/api-docs.yaml
openapi: 3.0.0
info:
  title: HMS Backend Code Wizards
  version: 1.0.0
  description:  Description of all the paths of the backend as well as needed information and status codes
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/assignment/{assignCode}:
    get:
      summary: Get details for an assignment
      operationId: getAssignments
      tags:
        - Assignment
      responses:
        '200':
          description: Details for assignment successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Assignment'
                type: object
        '404':
          description: Assignments not found
        '500':
          description: Internal server error
...
  /api/submission/submit:
    post:
      summary: Submit an assignment
      operationId: submitAssignment
      security:
        - bearerAuth: []
      tags:
        - Submission
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                assignCode:
                  type: string
      responses:
        '201':
          description: Submission submitted successfully
        '404':
          description: User/Assignment not found
        '400':
          description: Bad request
        '401':
          description: Unauthorized - Missing or invalid token
        '500':
          description: Internal server error
  /api/submission/assignment/{assignCode}:
    get:
      summary: Get submissions for an assignment
      operationId: getSubmissionsByAssignment
      tags:
        - Submission
      parameters:
        - name: assignCode
          in: path
          required: true
          description: The assignment code
          schema:
            type: string
      responses:
        '200':
          description: Details for assignment successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Submission'
                type: array
        '404':
          description: Submissions not found
        '500':
          description: Internal server error
  /api/submission/{username}/{assignCode}:
    get:
      summary: Get a specific submission by username and assignment code
      operationId: getSubmissionByUserAndAssignment
      tags:
        - Submission
      parameters:
        - name: username
          in: path
          required: true
          description: The username of the student
          schema:
            type: string
        - name: assignCode
          in: path
          required: true
          description: The assignment code
          schema:
            type: string
      responses:
        '200':
          description: Details for assignment successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Submission'
                type: object
        '400':
          description: Bad request
        '404':
          description: Submissions not found
        '500':
          description: Internal server error
  /api/submission/grade/{username}/{assignCode}:
    put:
      summary: Grade a submission
      operationId: gradeSubmission
      security:
        - bearerAuth: []
      tags:
        - Submission
      parameters:
        - name: username
          in: path
          required: true
          description: The username of the student
          schema:
            type: string
        - name: assignCode
          in: path
          required: true
          description: The assignment code
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                grade:
                  type: number
                  minimum: 0
                  maximum: 250
                  description: Max is 250 unless the assignment counts less
                feedback:
                  type: string
      responses:
        '200':
          description: Submission graded successfully
        '400':
          description: Bad request
        '404':
          description: Submission or assignment not found
        '500':
          description: Internal server error
  /api/submission/delete/{id}:
    delete:
      summary: Delete a submission
      operationId: deleteSubmission
      security:
        - bearerAuth: []
      tags:
        - Submission
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the submission to delete
          schema:
            type: string
      responses:
        '200':
          description: Submission deleted successfully
        '400':
          description: Bad request
        '404':
          description: Submission not found
        '500':
          description: Internal server error
...